- Move ItemTypeIcon to icons path.
