+++ app/src/main/java/com/shov/unlimstorage/views/files/FileInfoScreen.kt
@@ -24,15 +24,16 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.google.accompanist.insets.navigationBarsPadding
 import com.shov.unlimstorage.R
-import com.shov.unlimstorage.ui.*
+import com.shov.unlimstorage.ui.UserInfo
 import com.shov.unlimstorage.ui.buttons.CustomIconButton
+import com.shov.unlimstorage.ui.icons.ItemTypeIcon
 import com.shov.unlimstorage.ui.texts.CustomText
 import com.shov.unlimstorage.ui.texts.TextInfo
 import com.shov.unlimstorage.utils.converters.toPrettyString
 import com.shov.unlimstorage.utils.observeConnectivityAsFlow
 import com.shov.unlimstorage.viewModels.DownloadViewModel
-import com.shov.unlimstorage.viewModels.common.TopAppBarViewModel
 import com.shov.unlimstorage.viewModels.common.ScaffoldViewModel
+import com.shov.unlimstorage.viewModels.common.TopAppBarViewModel
 import com.shov.unlimstorage.viewModels.files.FileInfoViewModel
 import com.shov.unlimstorage.viewModels.provider.singletonViewModel
 import com.shov.unlimstorage.viewStates.FileInfoState
